% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignFormula.R
\name{DesignFormula}
\alias{DesignFormula}
\title{Internal function extracting design matrices from formulas in the DALSM function and computing penalty related matrices.}
\usage{
DesignFormula(
  formula,
  data,
  K = 10,
  pen.order = 2,
  knots.x = NULL,
  n = NULL,
  nointercept = FALSE
)
}
\description{
Internal function extracting design matrices from formulas in the DALSM function and computing penalty related matrices.
}
\keyword{internal}
