% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freehmsData.R
\docType{data}
\name{freehmsData}
\alias{freehmsData}
\title{Perception of gay men and lesbians in Wallonia, Belgium.}
\format{
A data frame with 552 rows and 4 columns.
\describe{
 \item{\code{freehms}}{Ordinal response (1: Agree strongly to 5: Disagree strongly).}
 \item{\code{age}}{Age of the respondent.}
 \item{\code{gndr}}{Gender of the respondent.}
 \item{\code{eduyrs}}{Number of years of education completed.}
}
}
\usage{
data(freehmsData)
}
\description{
Data extracted from the European Social Survey (2018) conducted in
a series of European countries including Belgium and one of its three regions, Wallonia.
Each of the 552 participants (aged at least 15) was asked to react to the following statement,
"Gay men and lesbians should be free to live their own life as they wish",
with a positioning on a Likert scale going from 1 (=Agree strongly) to 5 (=Disagree strongly),
with 3 labelled as Neither agree nor disagree.
#'
}
\references{
Lambert, P. and Gressani, 0. (2023) Penalty parameter selection and asymmetry corrections
to Laplace approximations in Bayesian P-splines models. Statistical Modelling. doi:10.1177/1471082X231181173. Preprint: arXiv:2210.01668.

European Social Survey Round 9 Data (2018)
 Data file edition 3.1. NSD - Norwegian Centre for Research Data, Norway.
}
\keyword{datasets}
