% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordgam.R
\name{ordgam}
\alias{ordgam}
\title{Fit a proportional odds model for ordinal data specified using a formula with (potentially) additive terms (for given fixed penalty parameters).}
\usage{
ordgam(
  formula,
  data,
  nc = NULL,
  K = 10,
  pen.order = 2,
  descending = TRUE,
  Wood.test = TRUE,
  theta0 = NULL,
  lambda0 = NULL,
  ci.level = 0.95
)
}
\arguments{
\item{formula}{A model formula}

\item{data}{A data frame containing a column 'y' with the ordinal response (taking integer values) besides the covariates.}

\item{nc}{(optional) Number of categories for the ordinal response.}

\item{K}{Number of B-splines to model each additive term (Default: 10).}

\item{pen.order}{Penalty order (Default: 2).}

\item{descending}{Logical indicating if the odds of the response taking a value in the upper scale should be preferred over values in the lower scale.}

\item{Wood.test}{Logical indicating if the significance test described by S. Wood (Biometrika 2013)  should be used to test for additive term significance. When FALSE, an approximate Chi2 test based on the effective degrees of freedom for each additive term is reported.}

\item{theta0}{(Optional) Vector containing starting values for the regression parameters}

\item{lambda0}{Vector of penalty parameters for the additive terms (Default: 10 for each additive term).}

\item{ci.level}{Confidence levels of the computed credible intervals for the regression parameters.}
}
\value{
an object of type \code{\link{ordgam.object}}.
}
\description{
Fit a proportional odds model for ordinal data specified using a formula with (potentially) additive terms (for given fixed penalty parameters).
}
\examples{
library(ordgam)
data(freehmsData)
mod = ordgam(freehms ~ gndr + s(eduyrs) + s(age),
             data=freehmsData, descending=TRUE)
print(mod)
plot(mod)

}
\references{
Lambert, P. and Gressani, 0. (2022) Penalty parameter selection and
asymmetry corrections to Laplace approximations in Bayesian P-splines models.
arXiv:2210.01668.
}
\seealso{
\code{\link{ordregr}}, \code{\link{ordgam.object}}.
}
